final changes in jaccard index, fixes in segmentation examples, functionality for fullpreprocessing and full segmentation, and accuracy example notebook
